karuasa/requirements.txt
# Core Django
Django==4.2.7

# Image Processing
Pillow>=10.0.0

# Forms and UI
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-ckeditor==6.7.0

# Admin Theme
django-jazzmin==2.6.0

# Database (MySQL for local, SQLite built-in for Render)
pymysql==1.1.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# AI Integration
google-generativeai==0.3.2

# Production Server
gunicorn==21.2.0

# Static Files for Production
whitenoise==6.6.0
